// client/src/bindings/UserTransaction.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UserTransactionStatus } from "./UserTransactionStatus";

export type UserTransaction = { id: number, user_id: number, plan_id: number, amount: number, status: UserTransactionStatus, stripe_payment_intent_id: string | null, created_at: string, updated_at: string, };