// client/src/bindings/UserTransactionStatus.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type UserTransactionStatus = "Pending" | "Canceled" | "Processing" | "RequiresAction" | "RequiresCapture" | "RequiresConfirmation" | "RequiresPaymentMethod" | "Succeeded" | "Refunded";