// client/src/bindings/Plan.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type Plan = { id: number, enabled: boolean, name: string, price: string, data_limit: number, duration_days: number, description: string | null, created_at: string, updated_at: string, };